Make circles be in line, and simplify halftone() code
It seems that circles in the same row didn't have their centres in
line with each other - it was more like the tops of the circles in
a row were in line.

Now the centres are in line.

I've also re-worked the code in the loop in the `halftone()` method
so that it (hopefully) makes more sense as to what it's doing,
and added comments.